Move logic to ClustersService
