+++ src/ocm/components/clusterDetail/ClusterDetailsButtonGroup.tsx
@@ -1,21 +1,35 @@
-import { Flex, FlexItem, Button, ButtonVariant } from '@patternfly/react-core';
 import React from 'react';
+import { Flex, FlexItem, Button, ButtonVariant } from '@patternfly/react-core';
 import { Cluster, canDownloadClusterLogs, useAlerts, KubeconfigDownload } from '../../../common';
 import { downloadClusterInstallationLogs } from './utils';
 import { useModalDialogsContext } from '../hosts/ModalDialogsContext';
 import { canAbortInstallation } from '../clusters/utils';
 import { onFetchEvents } from '../fetching/fetchEvents';
 import ViewClusterEventsButton from '../../../common/components/ui/ViewClusterEventsButton';
-
-type ClusterDetailsButtonGroupProps = {
-  cluster: Cluster;
-};
+import { handleApiError } from '../../api';
+import { getApiErrorMessage } from '../../../common/utils';
+import { useTranslation } from '../../../common/hooks/use-translation-wrapper';
+import ClustersService from '../../services/ClustersService';
 
 const getID = (suffix: string) => `cluster-detail-${suffix}`;
 
-const ClusterDetailsButtonGroup: React.FC<ClusterDetailsButtonGroupProps> = ({ cluster }) => {
+const ClusterDetailsButtonGroup = ({ cluster }: { cluster: Cluster }) => {
   const { addAlert } = useAlerts();
+  const { t } = useTranslation();
   const { cancelInstallationDialog } = useModalDialogsContext();
+
+  const handleDownload = React.useCallback(async () => {
+    try {
+      await ClustersService.downloadKubeconfigFile(cluster.id);
+    } catch (e) {
+      handleApiError(e, (e) => {
+        addAlert({
+          title: t('ai:Could not download kubeconfig'),
+          message: getApiErrorMessage(e),
+        });
+      });
+    }
+  }, [addAlert, cluster.id, t]);
 
   return (
     <Flex
@@ -37,8 +51,8 @@ const ClusterDetailsButtonGroup: React.FC<ClusterDetailsButtonGroupProps> = ({ c
       <FlexItem>
         <KubeconfigDownload
           status={cluster.status}
-          clusterId={cluster.id}
           data-testid={getID('button-download-kubeconfig')}
+          handleDownload={handleDownload}
         />
       </FlexItem>
       <FlexItem>
@@ -48,7 +62,7 @@ const ClusterDetailsButtonGroup: React.FC<ClusterDetailsButtonGroupProps> = ({ c
         <Button
           data-testid="cluster-installation-logs-button"
           variant={ButtonVariant.link}
-          onClick={() => downloadClusterInstallationLogs(addAlert, cluster.id)}
+          onClick={() => void downloadClusterInstallationLogs(addAlert, cluster.id)}
           isDisabled={!canDownloadClusterLogs(cluster)}
         >
           Download Installation Logs
